+++ src/views/components/l13-diff-actions/l13-diff-actions.component.ts
@@ -40,6 +40,9 @@ export class L13DiffActionsComponent extends L13Element<L13DiffActionsViewModel>
 	@L13Query('#l13_select_untracked')
 	public selectUntracked:HTMLElement;
 	
+	@L13Query('#l13_select_all')
+	public selectAll:HTMLElement;
+	
 	@L13Query('#l13_copy_left')
 	public copyLeft:HTMLElement;
 	
@@ -53,11 +56,13 @@ export class L13DiffActionsComponent extends L13Element<L13DiffActionsViewModel>
 		setLabelText(this.selectDeleted, 'Select all deleted files');
 		setLabelText(this.selectModified, 'Select all modfied files');
 		setLabelText(this.selectUntracked, 'Select all untracked files');
+		setLabelText(this.selectAll, 'Select all files');
 		setLabelText(this.copyLeft, 'Copy selection to the right folder');
 		
 		this.selectDeleted.addEventListener('click', ({ metaKey, ctrlKey }) => this.list.selectByStatus('deleted', isMetaKey(ctrlKey, metaKey)));
 		this.selectModified.addEventListener('click', ({ metaKey, ctrlKey }) => this.list.selectByStatus('modified', isMetaKey(ctrlKey, metaKey)));
 		this.selectUntracked.addEventListener('click', ({ metaKey, ctrlKey }) => this.list.selectByStatus('untracked', isMetaKey(ctrlKey, metaKey)));
+		this.selectAll.addEventListener('click', () => this.list.selectAll());
 		
 		this.copyLeft.addEventListener('click', () => this.list.copy('left'));
 		this.copyRight.addEventListener('click', () => this.list.copy('right'));