+++ src/views/main.ts
@@ -18,9 +18,17 @@ import './components/l13-diff-input/l13-diff-input.component';
 import './components/l13-diff-input/l13-diff-input.service';
 import './components/l13-diff-input/l13-diff-input.viewmodel';
 
+import './components/l13-diff-intro/l13-diff-intro.component';
+import './components/l13-diff-intro/l13-diff-intro.service';
+import './components/l13-diff-intro/l13-diff-intro.viewmodel';
+
 import './components/l13-diff-list/l13-diff-list.component';
 import './components/l13-diff-list/l13-diff-list.service';
 import './components/l13-diff-list/l13-diff-list.viewmodel';
+
+import './components/l13-diff-search/l13-diff-search.component';
+import './components/l13-diff-search/l13-diff-search.service';
+import './components/l13-diff-search/l13-diff-search.viewmodel';
 
 import './components/l13-diff-menu/l13-diff-menu.component';
 import './components/l13-diff-menu/l13-diff-menu.service';