+++ CHANGELOG.md
@@ -1,10 +1,22 @@
 # Change Log
 All notable changes to the "L13 Diff" extension will be documented in this file.
 
+## [0.15.0] - 2019-06-09
+
+### Added
+- Added search widget (`Ctrl/Cmd + F`) to filter diff result.
+- Added select all in list view with keyboard shortcut `Ctrl/Cmd + A` or button.
+- Added `Alt + Double Click` to open diff or file to side.
+- Added start screen for keyboard shortcuts.
+
+### Fixed
+- Fixed menu visiblity if path was selected and input was clicked again, but menu did not appear.
+- Fixed auto link feature of markdown in changelog for VS Code Marketplace.
+
 ## [0.14.0] - 2019-05-26
 
 ### Added
-- Added keyboard support for selection in list view. Please see README.md
+- Added keyboard support for selecting items in the list view.
 
 ### Fixed
 - Fixed a bug if an item was select in the menu with enter, but value was not set.